Simplify some devices, putting them back in global event base
The i8253 and mc146818 don't really need dedicated threads, so let
them be a part of the global event base again. This cuts down on
complexity a bit and hopefully will help squash some issues with the
corruption of the i8253 I'm seeing in a 9front vm (but so far am not
seeing elsewhere).