more thread safety, more fun
This adds in a bunch of hacks to protect the event bases at all
costs. Super messy, but a good start.

Still some edge cases that don't seem solved in the current vmd(8)
anyway with respect to pausing while a guest is booting. Seems to not
work well yet.

Virtio isn't yet protected fully but doesn't seem really exposed as it
doesn't seem to deal with timers other than with some vmmci
stuff. Still needed, but lower priority.