Standardize reading messages off the pipes
Also clean up the fatal error messages.

Right now, vm_pipe_read() will return some value, but there's no value
for an invalid read as the expectation is the program exits. This
probably isn't the best...might need to switch to int8_t instead of
uint8_t and use -1 as an error since 0 is taken for a message already.